+++ meteor/client/ui/SegmentTimeline/SegmentTimelineContainer.tsx
@@ -31,8 +31,6 @@ import { Rundown, RundownId, Rundowns } from '../../../lib/collections/Rundowns'
 import { PartInstanceId, PartInstances, PartInstance } from '../../../lib/collections/PartInstances'
 import { PieceInstances } from '../../../lib/collections/PieceInstances'
 import { Parts, PartId, Part } from '../../../lib/collections/Parts'
-import { doUserAction, UserAction } from '../../lib/userAction'
-import { MeteorCall } from '../../../lib/api/methods'
 import { Tracker } from 'meteor/tracker'
 import { Meteor } from 'meteor/meteor'
 import RundownViewEventBus, {
@@ -510,15 +508,6 @@ export const SegmentTimelineContainer = translateWithTracker<IProps, IState, ITr
 					this.onRewindSegment()
 					this.onShowEntireSegment()
 				}
-
-				if (this.props.segmentui && this.props.segmentui.orphaned) {
-					const { t } = this.props
-					// TODO: This doesn't seem right? componentDidUpdate can be triggered in a lot of different ways.
-					// What is this supposed to do?
-					doUserAction(t, undefined, UserAction.RESYNC_SEGMENT, () =>
-						MeteorCall.userAction.resyncSegment('', this.props.segmentui!.rundownId, this.props.segmentui!._id)
-					)
-				}
 			}
 			if (
 				// the segment isn't live, is next, and the nextPartId has changed
@@ -771,8 +760,8 @@ export const SegmentTimelineContainer = translateWithTracker<IProps, IState, ITr
 				const currentLivePart = currentLivePartInstance.part
 
 				const partOffset =
-					this.context.durations?.partDisplayStartsAt?.[unprotectString(currentLivePart._id)] -
-						this.context.durations.partDisplayStartsAt[unprotectString(this.props.parts[0].instance.part._id)] || 0
+					(this.context.durations?.partDisplayStartsAt?.[unprotectString(currentLivePart._id)] || 0) -
+					(this.context.durations?.partDisplayStartsAt?.[unprotectString(this.props.parts[0]?.instance.part._id)] || 0)
 
 				let isExpectedToPlay = !!currentLivePartInstance.timings?.startedPlayback
 				const lastTake = currentLivePartInstance.timings?.take
@@ -781,7 +770,7 @@ export const SegmentTimelineContainer = translateWithTracker<IProps, IState, ITr
 				const virtualStartedPlayback =
 					(lastTake || 0) > (lastStartedPlayback || -1)
 						? lastTake
-						: lastStartedPlayback
+						: lastStartedPlayback !== undefined
 						? lastStartedPlayback - lastTakeOffset
 						: undefined
 
@@ -810,10 +799,6 @@ export const SegmentTimelineContainer = translateWithTracker<IProps, IState, ITr
 			} else {
 				this.isVisible = true
 			}
-		}
-
-		convertTimeToPixels = (time: number) => {
-			return Math.round(this.props.timeScale * time)
 		}
 
 		startLive = () => {