+++ meteor/lib/collections/PackageContainerPackageStatus.ts
@@ -4,6 +4,7 @@ import { StudioId } from './Studios'
 import { registerIndex } from '../database'
 import { ExpectedPackageStatusAPI } from '@sofie-automation/blueprints-integration'
 import { ExpectedPackageId } from './ExpectedPackages'
+import { PeripheralDeviceId } from './PeripheralDevices'
 
 /**
  * The PackageContainerPackageStatuses-collection contains statuses about "a Package on a specific PackageContainer"
@@ -29,6 +30,9 @@ export interface PackageContainerPackageStatusDB {
 	/** The Package this status is for */
 	packageId: string
 
+	/** Which PeripheralDevice this update came from */
+	deviceId: PeripheralDeviceId
+
 	/** The status of the Package */
 	status: ExpectedPackageStatusAPI.PackageContainerPackageStatus
 
@@ -39,18 +43,21 @@ export const PackageContainerPackageStatuses = createMongoCollection<
 	PackageContainerPackageStatusDB,
 	PackageContainerPackageStatusDB
 >('packageContainerPackageStatuses')
-registerCollection('PackageContainerStatuses', PackageContainerPackageStatuses)
+registerCollection('PackageContainerPackageStatuses', PackageContainerPackageStatuses)
 
 registerIndex(PackageContainerPackageStatuses, {
 	studioId: 1,
 	containerId: 1,
 	packageId: 1,
 })
+registerIndex(PackageContainerPackageStatuses, {
+	deviceId: 1,
+})
 
 export function getPackageContainerPackageId(
 	studioId: StudioId,
 	containerId: string,
-	packageId: string | ExpectedPackageId
+	packageId: ExpectedPackageId
 ): PackageContainerPackageId {
 	return protectString(`${studioId}_${containerId}_${packageId}`)
 }