+++ meteor/server/api/ingest/actions.ts
@@ -11,7 +11,6 @@ import { logger } from '../../logging'
 import { RundownPlaylists, RundownPlaylistId } from '../../../lib/collections/RundownPlaylists'
 import { TriggerReloadDataResponse } from '../../../lib/api/userActions'
 import { waitForPromise, waitForPromiseAll } from '../../../lib/lib'
-import { Segment } from '../../../lib/collections/Segments'
 import { GenericDeviceActions } from './genericDevice/actions'
 import {
 	runPlayoutOperationWithLock,
@@ -43,19 +42,6 @@ export namespace IngestActions {
 		}
 	}
 
-	export function reloadSegment(rundown: Rundown, segment: Segment): TriggerReloadDataResponse {
-		const device = getPeripheralDeviceFromRundown(rundown)
-
-		if (device.type === PeripheralDeviceAPI.DeviceType.MOS) {
-			// MOS doesn't support reloading a segment, so do the whole rundown
-			return MOSDeviceActions.reloadRundown(device, rundown)
-		} else if (device.type === PeripheralDeviceAPI.DeviceType.INEWS) {
-			return GenericDeviceActions.reloadSegment(device, rundown, segment)
-		} else {
-			throw new Meteor.Error(400, `The device ${device._id} does not support the method "reloadRundown"`)
-		}
-	}
-
 	/**
 	 * Notify the device on what part is currently playing
 	 * @param rundown
@@ -72,6 +58,8 @@ export namespace IngestActions {
 		if (!playlist) throw new Meteor.Error(501, `Orphaned rundown: "${rundown._id}"`)
 		if (playlist.rehearsal) currentPlayingPart = null
 
+		const previousPlayingPartExternalId: string | null = rundown.notifiedCurrentPlayingPartExternalId || null
+
 		const currentPlayingPartExternalId: string | null = currentPlayingPart ? currentPlayingPart.externalId : null
 		if (currentPlayingPartExternalId) {
 			Rundowns.update(this._id, {
@@ -83,7 +71,7 @@ export namespace IngestActions {
 		} else {
 			Rundowns.update(this._id, {
 				$unset: {
-					currentPlayingStoryStatus: 1,
+					notifiedCurrentPlayingPartExternalId: 1,
 				},
 			})
 			delete rundown.notifiedCurrentPlayingPartExternalId
@@ -96,7 +84,7 @@ export namespace IngestActions {
 			MOSDeviceActions.notifyCurrentPlayingPart(
 				device,
 				rundown,
-				rundown.notifiedCurrentPlayingPartExternalId || null,
+				previousPlayingPartExternalId,
 				currentPlayingPartExternalId
 			)
 		}